current/pyproject.toml
[project]
name = "youtube-ia-archiver"
version = "0.1.0"
description = "DaGhE Module: Archive YouTube playlists to Internet Archive"
requires-python = ">=3.11"
dependencies = [
    "yt-dlp>=2026.3.13",
    "yt-dlp-ejs",
    "pyyaml>=6.0.1",
    "internetarchive>=5.4.2"
]

[tool.uv]
package = false
